+++ src/main/platform.c
@@ -1787,7 +1787,7 @@ SEXP attribute_hidden do_Cstack_info(SEXP call, SEXP op, SEXP args, SEXP rho)
     PROTECT(nms = allocVector(STRSXP, 4));
     INTEGER(ans)[0] = (R_CStackLimit == -1) ? NA_INTEGER : R_CStackLimit;
     INTEGER(ans)[1] = (R_CStackLimit == -1) ? NA_INTEGER :
-	R_CStackDir * (R_CStackStart - (unsigned long) &ans);
+	R_CStackDir * (R_CStackStart - (uintptr_t) &ans);
     INTEGER(ans)[2] = R_CStackDir;
     INTEGER(ans)[3] = R_EvalDepth;
     SET_STRING_ELT(nms, 0, mkChar("size"));
